Refactor item handling: remove unused & deprecated object manager injections, introduce common line items
